Apply a function over (multi-band) pixels

This generic function applies a function on pixels of a data cube, an R array, or other classes if implemented.

apply_pixel(x, ...)

Arguments

x
input data

...
additional arguments passed to method implementations

Value

return value and type depend on the class of x

See also

Examples

# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db"))
}

#> Error in libgdalcubes_create_image_collection_from_format(files, format,     out_file, unroll_archives): ERROR in collection_format::load_file(): image collection format file does not exist.

v = cube_view(extent=list(left=388941.2, right=766552.4,
              bottom=4345299, top=4744931, t0="2018-04", t1="2018-06"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")

L8.col = image_collection(file.path(tempdir(), "L8.db"))

#> Error in image_collection(file.path(tempdir(), "L8.db")): file.exists(path) is not TRUE

apply_pixel(raster_cube(L8.col, v), "(B05-B04)/(B05+B04)", "NDVI")

#> A GDAL data cube proxy object
#> 
#> Dimensions:
#>        low     high count       pixel_size chunk_size
#> t  2018-04  2018-06     3              P1M         16
#> y  4345299  4744931   526 759.756653992395        256
#> x 388941.2 766552.4   497 759.781086519115        256
#> 
#> Bands:
#>   name offset scale nodata unit
#> 1 NDVI      0     1    NaN     
#> 

d <- c(4,16,128,128)
x <- array(rnorm(prod(d)), d)
y <- apply_pixel(x, function(v) {
  v[1] + v[2] + v[3] - v[4]
})