Query data cube timeseries at irregular spatial points

This function will overlay provided spatial points with a data cube and return time series of all bands of the cells for each query point, as a list of data.frame (one data frame per band) where rows correspond to points and columns represent time. If needed, point coordinates are automatically transformed to the SRS of the data cube.

query_timeseries(x, px, py, srs)

Arguments

x
source data cube

px
vector of x coordinates

py
vector of y coordinates

srs
spatial reference system string identifer (as GDAL understands)

Value

a list of data.frames (one per band / variable) with one row per point and one column per data cube time slice

Examples

# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db"))
}
L8.col = image_collection(file.path(tempdir(), "L8.db"))
v = cube_view(extent=list(left=388941.2, right=766552.4,
              bottom=4345299, top=4744931, t0="2018-01-01", t1="2018-12-02"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P14D")
L8.cube = raster_cube(L8.col, v)
L8.rgb = select_bands(L8.cube, c("B02", "B03", "B04"))

x = seq(from = 388941.2, to = 766552.4, length.out = 10)
y = seq(from = 4345299, to = 4744931, length.out = 10)

query_timeseries(L8.rgb, x, y, srs(L8.rgb))

#> $B02
#>    2018-01-01 2018-01-15 2018-01-29 2018-02-12 2018-02-26 2018-03-12 2018-03-26
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN        NaN       8678        NaN        NaN
#> 3         NaN        NaN        NaN        NaN      12592        NaN        NaN
#> 4         NaN        NaN        NaN        NaN      30430        NaN        NaN
#> 5       11077        NaN        NaN        NaN      28223        NaN       9425
#> 6       10125        NaN       7754        NaN        NaN        NaN       9851
#> 7        8921        NaN        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-04-09 2018-04-23 2018-05-07 2018-05-21 2018-06-04 2018-06-18 2018-07-02
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN       9869        NaN        NaN       9835        NaN        NaN
#> 3         NaN      10166        NaN        NaN       9797        NaN        NaN
#> 4         NaN      23018        NaN        NaN       9340        NaN        NaN
#> 5        9188      26634        NaN      32825      19620        NaN        NaN
#> 6        9576       9296        NaN       9369        NaN        NaN       8875
#> 7         NaN       9514        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-07-16 2018-07-30 2018-08-13 2018-08-27 2018-09-10 2018-09-24 2018-10-08
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN       9067        NaN        NaN        NaN
#> 3         NaN        NaN        NaN       8879        NaN        NaN        NaN
#> 4         NaN        NaN        NaN       8595        NaN        NaN        NaN
#> 5         NaN        NaN        NaN       8665        NaN      10141        NaN
#> 6         NaN        NaN        NaN      12651        NaN       7796        NaN
#> 7         NaN        NaN        NaN       8232        NaN       7649        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-10-22 2018-11-05 2018-11-19
#> 1         NaN        NaN        NaN
#> 2         NaN        NaN       7710
#> 3         NaN        NaN       7426
#> 4         NaN        NaN       7940
#> 5        7546        NaN       8458
#> 6        7620        NaN       8327
#> 7         NaN        NaN        NaN
#> 8         NaN        NaN        NaN
#> 9         NaN        NaN        NaN
#> 10        NaN        NaN        NaN
#> 
#> $B03
#>    2018-01-01 2018-01-15 2018-01-29 2018-02-12 2018-02-26 2018-03-12 2018-03-26
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN        NaN       7656        NaN        NaN
#> 3         NaN        NaN        NaN        NaN      11964        NaN        NaN
#> 4         NaN        NaN        NaN        NaN      29462        NaN        NaN
#> 5       10487        NaN        NaN        NaN      27222        NaN       8061
#> 6        9313        NaN       6910        NaN        NaN        NaN       9000
#> 7        8101        NaN        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-04-09 2018-04-23 2018-05-07 2018-05-21 2018-06-04 2018-06-18 2018-07-02
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN       9159        NaN        NaN       9336        NaN        NaN
#> 3         NaN       9595        NaN        NaN       9977        NaN        NaN
#> 4         NaN      21931        NaN        NaN       9065        NaN        NaN
#> 5        7959      25409        NaN      31770      18918        NaN        NaN
#> 6        8873       8269        NaN       8774        NaN        NaN       8155
#> 7         NaN       8789        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-07-16 2018-07-30 2018-08-13 2018-08-27 2018-09-10 2018-09-24 2018-10-08
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN       8092        NaN        NaN        NaN
#> 3         NaN        NaN        NaN       8366        NaN        NaN        NaN
#> 4         NaN        NaN        NaN       7767        NaN        NaN        NaN
#> 5         NaN        NaN        NaN       7842        NaN       9627        NaN
#> 6         NaN        NaN        NaN      11427        NaN       7174        NaN
#> 7         NaN        NaN        NaN       7503        NaN       6981        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-10-22 2018-11-05 2018-11-19
#> 1         NaN        NaN        NaN
#> 2         NaN        NaN       7082
#> 3         NaN        NaN       6776
#> 4         NaN        NaN       7526
#> 5        6796        NaN       7652
#> 6        7038        NaN       7516
#> 7         NaN        NaN        NaN
#> 8         NaN        NaN        NaN
#> 9         NaN        NaN        NaN
#> 10        NaN        NaN        NaN
#> 
#> $B04
#>    2018-01-01 2018-01-15 2018-01-29 2018-02-12 2018-02-26 2018-03-12 2018-03-26
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN        NaN       7689        NaN        NaN
#> 3         NaN        NaN        NaN        NaN      12118        NaN        NaN
#> 4         NaN        NaN        NaN        NaN      31194        NaN        NaN
#> 5       10936        NaN        NaN        NaN      28889        NaN       7476
#> 6        9672        NaN       6816        NaN        NaN        NaN       9433
#> 7        7977        NaN        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-04-09 2018-04-23 2018-05-07 2018-05-21 2018-06-04 2018-06-18 2018-07-02
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN       8081        NaN        NaN       8395        NaN        NaN
#> 3         NaN       8891        NaN        NaN       9908        NaN        NaN
#> 4         NaN      23271        NaN        NaN       7610        NaN        NaN
#> 5        7453      26758        NaN      32998      19156        NaN        NaN
#> 6        9522       8054        NaN       7399        NaN        NaN       6851
#> 7         NaN       9235        NaN        NaN        NaN        NaN        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-07-16 2018-07-30 2018-08-13 2018-08-27 2018-09-10 2018-09-24 2018-10-08
#> 1         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 2         NaN        NaN        NaN       6859        NaN        NaN        NaN
#> 3         NaN        NaN        NaN       7032        NaN        NaN        NaN
#> 4         NaN        NaN        NaN       6623        NaN        NaN        NaN
#> 5         NaN        NaN        NaN       6979        NaN       9085        NaN
#> 6         NaN        NaN        NaN      10503        NaN       6309        NaN
#> 7         NaN        NaN        NaN       6541        NaN       6386        NaN
#> 8         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 9         NaN        NaN        NaN        NaN        NaN        NaN        NaN
#> 10        NaN        NaN        NaN        NaN        NaN        NaN        NaN
#>    2018-10-22 2018-11-05 2018-11-19
#> 1         NaN        NaN        NaN
#> 2         NaN        NaN       7041
#> 3         NaN        NaN       6506
#> 4         NaN        NaN       7765
#> 5        6310        NaN       7709
#> 6        6762        NaN       7356
#> 7         NaN        NaN        NaN
#> 8         NaN        NaN        NaN
#> 9         NaN        NaN        NaN
#> 10        NaN        NaN        NaN
#>