List predefined image collection formats

gdalcubes comes with some predefined collection formats e.g. to scan Sentinel 2 data. This function lists available formats including brief descriptions.

collection_formats(print = TRUE)

Arguments

print

logical; should available formats and their descriptions be printed nicely, defaults to TRUE

Value

data.frame with columns name and description where the former describes the unique identifier that can be used in create_image_collection and the latter gives a brief description of the format.

Details

Image collection formats define how individual files / GDAL datasets relate to an image collection, i.e., which bands they contain, to which image they belong, and how to derive aquisition date/time. They are described as a set of regular expressions in a JSON file and used by gdalcubes to extract this information from the paths and/or filenames.

Examples

collection_formats()
#>    CHIRPS_v2_0_daily_p05_tif | Image collection format for CHIRPS v 2.0 daily
#>                              | global precipitation dataset (0.05 degrees
#>                              | resolution) from GeoTIFFs, expects list of .tif
#>                              | or .tif.gz files as input. [TAGS: CHIRPS,
#>                              | precipitation]
#>  CHIRPS_v2_0_monthly_p05_tif | Image collection format for CHIRPS v 2.0 monthly
#>                              | global precipitation dataset (0.05 degrees
#>                              | resolution) from GeoTIFFs, expects list of .tif
#>                              | or .tif.gz files as input. [TAGS: CHIRPS,
#>                              | precipitation]
#>            ESA_CCI_SM_ACTIVE | Collection format for ESA CCI soil moisture
#>                              | active product (version 4.7) [TAGS: Soil
#>                              | Moisture, ESA, CCI]
#>           ESA_CCI_SM_PASSIVE | Collection format for ESA CCI soil moisture
#>                              | passive product (version 4.7) [TAGS: Soil
#>                              | Moisture, ESA, CCI]
#>    GPM_IMERG_3B_DAY_GIS_V06A | Collection format for daily
#>                              | IMERG_3B_DAY_GIS_V06A data [TAGS: Precipitation,
#>                              | GPM, IMERG]
#>                      L8_L1TP | Collection format for Landsat 8 Level 1 TP
#>                              | product [TAGS: Landsat, USGS, Level 1, NASA]
#>                        L8_SR | Collection format for Landsat 8 surface
#>                              | reflectance product [TAGS: Landsat, USGS, Level
#>                              | 2, NASA, surface reflectance]
#>                      MCD64A1 | Collection format for the MODIS MCD64A1 product
#>                              | (burned area by day) [TAGS: MODIS, Fire]
#>                      MxD09GA | Collection format for selected bands from the
#>                              | MODIS MxD09GA (Aqua and Terra) product [TAGS:
#>                              | MODIS, surface reflectance]
#>                      MxD10A2 | Collection format for selected bands from the
#>                              | MODIS MxD10A2 (Aqua and Terra) v006 Snow Cover
#>                              | product [TAGS: MODIS, Snow Cover]
#>                      MxD11A1 | Collection format for selected bands from the
#>                              | MODIS MxD11A2 (Aqua and Terra) v006 Land Surface
#>                              | Temperature product [TAGS: MODIS, LST]
#>                      MxD11A2 | Collection format for selected bands from the
#>                              | MODIS MxD11A2 (Aqua and Terra) v006 Land Surface
#>                              | Temperature product [TAGS: MODIS, LST]
#>                      MxD13A2 | Collection format for selected bands from the
#>                              | MODIS MxD13A2 (Aqua and Terra) product [TAGS:
#>                              | MODIS, VI, NDVI, EVI]
#>                      MxD13A3 | Collection format for selected bands from the
#>                              | MODIS MxD13A3 (Aqua and Terra) product [TAGS:
#>                              | MODIS, VI, NDVI, EVI]
#>                      MxD13Q1 | Collection format for selected bands from the
#>                              | MODIS MxD13Q1 (Aqua and Terra) product [TAGS:
#>                              | MODIS, VI, NDVI, EVI]
#>                      MxD14A2 | Collection format for the MODIS MxD14A2 (Aqua
#>                              | and Terra) product [TAGS: MODIS, Fire]
#> PlanetScope_3B_AnalyticMS_SR | Image collection format for PlanetScope 4-band
#>                              | scenes [TAGS: PlanetScope, BOA, Surface
#>                              | Reflectance]
#>             Sentinel1_IW_GRD | Image collection format for Sentinel 1 Level 1
#>                              | GRD data as downloaded from the Copernicus Open
#>                              | Access Hub, expects a list of file paths as
#>                              | input. The format works on original ZIP
#>                              | compressed as well as uncompressed imagery.
#>                              | [TAGS: Sentinel, Copernicus, ESA, SAR]
#>                Sentinel2_L1C | Image collection format for Sentinel 2 Level 1C
#>                              | data as downloaded from the Copernicus Open
#>                              | Access Hub, expects a list of file paths as
#>                              | input. The format works on original ZIP
#>                              | compressed as well as uncompressed imagery.
#>                              | [TAGS: Sentinel, Copernicus, ESA, TOA]
#>            Sentinel2_L1C_AWS | Image collection format for Sentinel 2 Level 1C
#>                              | data in AWS [TAGS: Sentinel, Copernicus, ESA,
#>                              | TOA]
#>                Sentinel2_L2A | Image collection format for Sentinel 2 Level 2A
#>                              | data as downloaded from the Copernicus Open
#>                              | Access Hub, expects a list of file paths as
#>                              | input. The format should work on original ZIP
#>                              | compressed as well as uncompressed imagery.
#>                              | [TAGS: Sentinel, Copernicus, ESA, BOA, Surface
#>                              | Reflectance]
#>          Sentinel2_L2A_THEIA | Image collection format for Sentinel 2 Level 2A
#>                              | data as downloaded from Theia. [TAGS: Sentinel,
#>                              | ESA, Flat Reflectance, Theia]