Coerce gdalcubes object into a stars object

The function materializes a data cube as a temporary netCDF file and loads the file with the stars package.

st_as_stars.cube(.x, ...)

Arguments

.x
data cube object to coerce

...
not used

Value

stars object

Examples

# \donttest{
# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db"))
}

L8.col = image_collection(file.path(tempdir(), "L8.db"))
v = cube_view(extent=list(left=388941.2, right=766552.4,
              bottom=4345299, top=4744931, t0="2018-04", t1="2018-04"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")
if(require("stars"))
  st_as_stars(select_bands(raster_cube(L8.col, v), c("B04", "B05")))

#> Loading required package: stars

#> Loading required package: abind

#> Loading required package: sf

#> Linking to GEOS 3.8.0, GDAL 3.0.4, PROJ 7.0.0

#> stars object with 3 dimensions and 2 attributes
#> attribute(s):
#>       B04             B05       
#>  Min.   : 6119   Min.   : 5564  
#>  1st Qu.: 6928   1st Qu.: 6411  
#>  Median : 8717   Median :13519  
#>  Mean   :11064   Mean   :14300  
#>  3rd Qu.:10620   3rd Qu.:17073  
#>  Max.   :53002   Max.   :57497  
#>  NA's   :99070   NA's   :99070  
#> dimension(s):
#>      from  to  offset    delta                refsys point
#> x       1 497  388941  759.781 WGS 84 / UTM zone 18N    NA
#> y       1 526 4744931 -759.757 WGS 84 / UTM zone 18N    NA
#> time    1   1      NA       NA               POSIXct FALSE
#>                       values    
#> x                       NULL [x]
#> y                       NULL [y]
#> time [2018-04-01,2018-05-01)    

# }