Export a data cube as netCDF file(s)

This function will read chunks of a data cube and write them to a single (the default) or multitple (if chunked = TRUE) netCDF file(s). The resulting file(s) uses the enhanced netCDF-4 format, supporting chunking and compression.

write_ncdf(
  x,
  fname = tempfile(pattern = "gdalcubes", fileext = ".nc"),
  overwrite = FALSE,
  write_json_descr = FALSE,
  with_VRT = FALSE,
  pack = NULL,
  chunked = FALSE
)

Arguments

x
a data cube proxy object (class cube)

fname
output file name

overwrite
logical; overwrite output file if it already exists

write_json_descr
logical; write a JSON description of x as additional file

with_VRT
logical; write additional VRT datasets (one per time slice)

pack
reduce output file size by packing values (see Details), defaults to no packing

chunked
logical; if TRUE, write one netCDF file per chunk; defaults to FALSE

Value

returns (invisibly) the path of the created netCDF file(s)

Details

The resulting netCDF file(s) contain three dimensions (t, y, x) and bands as variables.

If write_json_descr is TRUE, the function will write an addition file with the same name as the NetCDF file but ".json" suffix. This file includes a serialized description of the input data cube, including all chained data cube operations.

To reduce the size of created files, values can be packed by applying a scale factor and an offset value and using a smaller integer data type for storage (only supported if chunked = TRUE). The pack argument can be either NULL (the default), or a list with elements type, scale, offset, and nodata. type can be any of "uint8", "uint16" , "uint32", "int16", or "int32". scale, offset, and nodata must be numeric vectors with length one or length equal to the number of data cube bands (to use different values for different bands). The helper function pack_minmax can be used to derive offset and scale values with maximum precision from minimum and maximum data values on original scale.

If chunked = TRUE, names of the produced files will start with name (with removed extension), followed by an underscore and the internal integer chunk number.

Note

Packing is currently ignored if chunked = TRUE

See also

Examples

# create image collection from example Landsat data only 
# if not already done in other examples
if (!file.exists(file.path(tempdir(), "L8.db"))) {
  L8_files <- list.files(system.file("L8NY18", package = "gdalcubes"),
                         ".TIF", recursive = TRUE, full.names = TRUE)
  create_image_collection(L8_files, "L8_L1TP", file.path(tempdir(), "L8.db"))
}

#> A GDAL image collection object, referencing 19 images with 12  bands
#> Images:
#>                                       name      left      top   bottom
#> 1 LC08_L1TP_013032_20180131_20180207_01_T1 -74.67898 41.39099 39.25027
#> 2 LC08_L1TP_013032_20180405_20180417_01_T1 -74.70333 41.39106 39.25080
#> 3 LC08_L1TP_013032_20180421_20180502_01_T1 -74.70681 41.39107 39.25098
#> 4 LC08_L1TP_013032_20180710_20180717_01_T1 -74.66854 41.39096 39.24991
#> 5 LC08_L1TP_013032_20180827_20180911_01_T1 -74.67202 41.39097 39.25000
#> 6 LC08_L1TP_013032_20181030_20181115_01_T1 -74.69637 41.39104 39.25062
#>       right            datetime
#> 1 -71.92546 2018-01-31T00:00:00
#> 2 -71.94695 2018-04-05T00:00:00
#> 3 -71.95411 2018-04-21T00:00:00
#> 4 -71.91114 2018-07-10T00:00:00
#> 5 -71.91472 2018-08-27T00:00:00
#> 6 -71.93979 2018-10-30T00:00:00
#>                                                 srs
#> 1 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> 2 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> 3 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> 4 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> 5 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> 6 +proj=utm +zone=18 +datum=WGS84 +units=m +no_defs
#> [ omitted 13 images ] 
#> 
#> Bands:
#>    name offset scale unit   nodata image_count
#> 1   B01      0     1      0.000000          19
#> 2   B02      0     1      0.000000          19
#> 3   B03      0     1      0.000000          19
#> 4   B04      0     1      0.000000          19
#> 5   B05      0     1      0.000000          19
#> 6   B06      0     1      0.000000          19
#> 7   B07      0     1      0.000000          19
#> 8   B08      0     1      0.000000          19
#> 9   B09      0     1      0.000000          19
#> 10  B10      0     1      0.000000          19
#> 11  B11      0     1      0.000000          19
#> 12  BQA      0     1                        19
#> 

L8.col = image_collection(file.path(tempdir(), "L8.db"))
v = cube_view(extent=list(left=388941.2, right=766552.4,
              bottom=4345299, top=4744931, t0="2018-04", t1="2018-04"),
              srs="EPSG:32618", nx = 497, ny=526, dt="P1M")
write_ncdf(select_bands(raster_cube(L8.col, v), c("B04", "B05")), fname=tempfile(fileext = ".nc"))